Moved all code to /src
